requirements.txt
# Core ML/AI
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# LLM & Transformers
transformers>=4.35.0
accelerate>=0.24.0

# RAG & Embeddings
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
sentence-transformers>=2.2.2

# DICOM
pydicom>=2.4.0
pillow>=10.0.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Frontend
streamlit>=1.28.0
plotly>=5.17.0

# Data Science
pandas>=2.1.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
requests>=2.31.0
